refactor (fe): rename admin-course-dashboard to index
